Split the mediator classes
