+++ synapse/src/main/java/synapse/converter/bir/MediatorConverters.java
@@ -15,66 +15,41 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package synapse.converter;
+package synapse.converter.bir;
 
 import common.BallerinaModel.Expression;
 import common.BallerinaModel.Expression.BallerinaExpression;
 import common.BallerinaModel.Expression.StringConstant;
 import common.BallerinaModel.Expression.XMLTemplate;
-import common.BallerinaModel.Function;
-import common.BallerinaModel.Import;
-import common.BallerinaModel.Parameter;
-import common.BallerinaModel.Resource;
-import common.BallerinaModel.Service;
 import common.BallerinaModel.Statement;
-import common.BallerinaModel.TypeDesc;
 import common.BallerinaModel.TypeDesc.BuiltinType;
-import synapse.model.Synapse.Api;
+import synapse.converter.ConversionContext;
+import synapse.converter.ScopeContext;
 import synapse.model.Synapse.Kind;
 import synapse.model.Synapse.PayloadFactory;
 import synapse.model.Synapse.Property;
-import synapse.model.Synapse.Sequence;
 import synapse.model.Synapse.SequenceMediator;
 import synapse.model.Synapse.SynapseNode;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
 /**
- * Converts a single {@link SynapseNode} into its Ballerina BIR representation,
- * accumulating the result
- * in a context.
- *
- * <p>
- * There are two families, distinguished by the context they operate on:
- * <ul>
- * <li><b>Root converters</b> ({@code BIRConverter<ConversionContext>}) handle
- * artifact-level elements
- * ({@code <api>}, {@code <sequence>}). They read the shared
- * {@link ConversionContext}, create a
- * {@link ScopeContext} for the body they generate, and write the result (a
- * service or function)
- * back to the shared context.</li>
- * <li><b>Mediator converters</b> ({@code BIRConverter<ScopeContext>}) handle
- * the mediators inside a
- * resource or sequence body, appending to the enclosing
- * {@link ScopeContext}.</li>
- * </ul>
- *
- * @param <C> the context this converter operates on
+ * The {@link BIRConverter} implementations for the mediators that appear inside a resource or sequence
+ * body, together with the dispatch that drives them. Each converter appends to the enclosing
+ * {@link ScopeContext}; {@link #convertMediators} walks a mediator list and routes each element to its
+ * converter, stopping at the first respond.
  */
-public interface BIRConverter<C> {
+public final class MediatorConverters {
 
-    Map<Kind, BIRConverter<ScopeContext>> MEDIATOR_CONVERTERS = Map.of(
+    private static final Map<Kind, BIRConverter<ScopeContext>> MEDIATOR_CONVERTERS = Map.of(
             Kind.PAYLOAD_FACTORY, new PayloadFactoryConverter(),
             Kind.PROPERTY, new PropertyConverter(),
             Kind.SEQUENCE_MEDIATOR, new SequenceMediatorConverter());
 
-    Import HTTP_IMPORT = new Import("ballerina", "http");
-
-    void convert(SynapseNode node, C context);
+    private MediatorConverters() {
+    }
 
     /**
      * Converts a list of Synapse mediators into Ballerina, accumulating the result
@@ -97,7 +72,7 @@ public interface BIRConverter<C> {
     static void convertMediators(List<SynapseNode> mediators, ScopeContext context) {
         for (SynapseNode mediator : mediators) {
             if (mediator.kind() == Kind.RESPOND) {
-                emitRespond(context);
+                context.emitRespond();
             } else {
                 BIRConverter<ScopeContext> converter = MEDIATOR_CONVERTERS.getOrDefault(mediator.kind(),
                         new UnsupportedConverter());
@@ -106,149 +81,6 @@ public interface BIRConverter<C> {
             if (context.isResponded()) {
                 break;
             }
-        }
-    }
-
-    private static void emitRespond(ScopeContext context) {
-        context.setResponded(true);
-        if (!context.isWithinResource()) {
-            return;
-        }
-        ensureResponseAvailable(context);
-        returnResponse(context);
-    }
-
-    private static void ensureResponseAvailable(ScopeContext context) {
-        context.importStatements().add(HTTP_IMPORT);
-        if (context.responseAvailable()) {
-            return;
-        }
-        if (context.isWithinResource()) {
-            context.statements().add(0, new Statement.BallerinaStatement("http:Response response = new;"));
-            context.setRespondInitialized(true);
-        } else {
-            context.setResponseParam(true);
-        }
-    }
-
-    private static void ensureContextAvailable(ScopeContext context) {
-        if (context.contextAvailable()) {
-            return;
-        }
-        if (context.isWithinResource()) {
-            context.statements().add(0, new Statement.BallerinaStatement("Context ctx = {};"));
-            context.setContextInitialized(true);
-        } else {
-            context.setContextParam(true);
-        }
-    }
-
-    private static void returnResponse(ScopeContext context) {
-        context.statements().add(new Statement.Return<>(Optional.of(new Expression.VariableReference("response"))));
-        context.setReturnType(new TypeDesc.BallerinaType("http:Response"));
-    }
-
-    /**
-     * Converts a Synapse {@code <api>} element into a Ballerina HTTP service.
-     */
-    class APIConverter implements BIRConverter<ConversionContext> {
-
-        private static final String DEFAULT_LISTENER_REF = "httpListener";
-        private static final String ROOT_RESOURCE_PATH = ".";
-
-        @Override
-        public void convert(SynapseNode node, ConversionContext context) {
-            Api api = (Api) node;
-            List<Resource> resources = new ArrayList<>();
-            for (SynapseNode child : api.resources()) {
-                if (child instanceof synapse.model.Synapse.Resource resource) {
-                    resources.add(convertResource(resource, context));
-                }
-            }
-            Service service = new Service(api.context(), DEFAULT_LISTENER_REF, resources);
-            context.addService(service);
-        }
-
-        private static Resource convertResource(synapse.model.Synapse.Resource resource, ConversionContext context) {
-            String method = resource.methods().toLowerCase();
-            String path = buildResourcePath(resource.path());
-
-            List<Parameter> parameters = new ArrayList<>();
-            for (String queryParam : resource.queryParams()) {
-                parameters.add(new Parameter(queryParam, BuiltinType.STRING));
-            }
-
-            ResourceContext resourceContext = new ResourceContext(context);
-            if (resource.inSequence() != null) {
-                convertMediators(resource.inSequence().mediators(), resourceContext);
-            }
-            context.addImports(ConversionContext.MAIN_BAL_FILE, resourceContext.importStatements());
-            TypeDesc returnType = resourceContext.returnType();
-            Optional<TypeDesc> returnTypeDesc = returnType == BuiltinType.NIL
-                    ? Optional.empty()
-                    : Optional.of(returnType);
-            return new Resource(method, path, parameters, returnTypeDesc, resourceContext.statements());
-        }
-
-        private static String buildResourcePath(String synapsePath) {
-            List<String> segments = new ArrayList<>();
-            for (String segment : synapsePath.split("/")) {
-                if (segment.isEmpty()) {
-                    continue;
-                }
-                if (segment.startsWith("{") && segment.endsWith("}")) {
-                    String paramName = segment.substring(1, segment.length() - 1);
-                    segments.add("[string " + paramName + "]");
-                } else {
-                    segments.add(segment);
-                }
-            }
-            return segments.isEmpty() ? ROOT_RESOURCE_PATH : String.join("/", segments);
-        }
-    }
-
-    /**
-     * Converts a top-level Synapse {@code <sequence>} into a Ballerina function
-     * whose body is the
-     * converted mediator flow. Whether the function takes an
-     * {@code http:Response response} parameter and
-     * whether it responds both fall out of converting the body rather than a
-     * separate pre-scan: when
-     * conversion reaches a {@code <payloadFactory>} (directly or down a call chain)
-     * the sequence takes a
-     * {@code response} parameter it mutates in place instead of returning, and an
-     * unreached payloadFactory
-     * (e.g. after a {@code <respond>}) leaves the function parameterless. The
-     * function's return type is
-     * likewise driven by the converted mediators: unless they yield a
-     * non-{@code nil} {@code returnType},
-     * the function is {@code nil}-returning.
-     */
-    class SequenceConverter implements BIRConverter<ConversionContext> {
-
-        @Override
-        public void convert(SynapseNode node, ConversionContext context) {
-            Sequence sequence = (Sequence) node;
-            SequenceContext sequenceContext = new SequenceContext(context);
-            convertMediators(sequence.mediators(), sequenceContext);
-            boolean containsPayloadFactory = sequenceContext.hasResponseParam();
-            boolean usesContext = sequenceContext.hasContextParam();
-            context.addSequenceMetadata(new ConversionContext.SequenceMetadata(
-                    sequence.name(), sequenceContext.isResponded(), containsPayloadFactory, usesContext));
-            context.addImports(ConversionContext.FUNCTIONS_BAL_FILE, sequenceContext.importStatements());
-            List<Parameter> params = new ArrayList<>();
-            if (usesContext) {
-                params.add(new Parameter("ctx", new TypeDesc.BallerinaType("Context")));
-            }
-            if (containsPayloadFactory) {
-                params.add(new Parameter("response", new TypeDesc.BallerinaType("http:Response")));
-            }
-            TypeDesc returnType = sequenceContext.returnType();
-            List<Statement> body = sequenceContext.statements();
-            Function function = returnType == BuiltinType.NIL
-                    ? new Function(sequence.name(), params, body)
-                    : new Function(sequence.name(), params, returnType, body);
-            context.addFunction(function);
         }
     }
 
@@ -263,7 +95,7 @@ public interface BIRConverter<C> {
      * drives a respond in
      * the enclosing scope.
      */
-    class SequenceMediatorConverter implements BIRConverter<ScopeContext> {
+    static class SequenceMediatorConverter implements BIRConverter<ScopeContext> {
 
         @Override
         public void convert(SynapseNode node, ScopeContext context) {
@@ -272,28 +104,28 @@ public interface BIRConverter<C> {
                     .orElse(null);
             List<Expression> args = new ArrayList<>();
             if (metadata != null && metadata.usesContext()) {
-                ensureContextAvailable(context);
+                context.ensureContextAvailable();
                 args.add(new Expression.VariableReference("ctx"));
             }
             if (metadata != null && metadata.containsPayloadFactory()) {
-                ensureResponseAvailable(context);
+                context.ensureResponseAvailable();
                 args.add(new Expression.VariableReference("response"));
             }
             context.statements().add(new Statement.CallStatement(
                     new Expression.FunctionCall(sequenceMediator.key(), args)));
             if (metadata != null && metadata.containsRespond()) {
-                emitRespond(context);
+                context.emitRespond();
             }
         }
     }
 
-    class PayloadFactoryConverter implements BIRConverter<ScopeContext> {
+    static class PayloadFactoryConverter implements BIRConverter<ScopeContext> {
 
         @Override
         public void convert(SynapseNode node, ScopeContext context) {
             PayloadFactory payloadFactory = (PayloadFactory) node;
             Expression value = extractValue(payloadFactory.mediaType(), payloadFactory.format());
-            ensureResponseAvailable(context);
+            context.ensureResponseAvailable();
             context.statements().add(new Statement.CallStatement(new Expression.MethodCall(
                     new Expression.VariableReference("response"), "setPayload", List.of(value))));
         }
@@ -318,7 +150,7 @@ public interface BIRConverter<C> {
      * converter therefore
      * first identifies its scope.
      */
-    class PropertyConverter implements BIRConverter<ScopeContext> {
+    static class PropertyConverter implements BIRConverter<ScopeContext> {
 
         private static final String TRANSPORT_SCOPE = "transport";
         private static final String AXIS2_SCOPE = "axis2";
@@ -362,7 +194,7 @@ public interface BIRConverter<C> {
         }
 
         private static void convertDefaultProperty(Property property, ScopeContext context) {
-            ensureContextAvailable(context);
+            context.ensureContextAvailable();
             if (REMOVE_ACTION.equals(property.action())) {
                 context.statements().add(new Statement.BallerinaStatement(
                         "ctx." + property.name() + " = " + BuiltinType.NIL + ";"));
@@ -383,7 +215,7 @@ public interface BIRConverter<C> {
         }
     }
 
-    class UnsupportedConverter implements BIRConverter<ScopeContext> {
+    static class UnsupportedConverter implements BIRConverter<ScopeContext> {
 
         @Override
         public void convert(SynapseNode node, ScopeContext context) {
